+++ tests/cli/test_cloud.py
@@ -13,13 +13,31 @@ from ploomber.telemetry.telemetry import DEFAULT_USER_CONF
 from ploomber import table
 
 
-@pytest.fixture()
+@pytest.fixture(autouse=True)
 def write_sample_conf(tmp_directory, monkeypatch):
+    """
+    Mocks the default location for the config file, and stores some
+    default content
+    """
+    # back up user config to prevent the tests from modifying it (the tests
+    # shouldn't access it directly but this is more robust)
+    user_config = Path(telemetry.DEFAULT_HOME_DIR, 'stats', DEFAULT_USER_CONF)
+    file_exists = user_config.exists()
+
+    if file_exists:
+        content = yaml.safe_load(user_config.read_text())
+
     monkeypatch.setattr(telemetry, 'DEFAULT_HOME_DIR', '.')
     stats = Path('stats')
     stats.mkdir()
     full_path = (stats / DEFAULT_USER_CONF)
     full_path.write_text("stats_enabled: False")
+
+    yield
+
+    if file_exists:
+        with user_config.open('w') as f:
+            yaml.safe_dump(content, f)
 
 
 def write_sample_pipeline(pipeline_id=None, status=None):
@@ -49,7 +67,7 @@ def get_tabular_pipeline(pipeline_id=None, verbose=None):
     return res.stdout
 
 
-def test_write_api_key(write_sample_conf):
+def test_write_api_key():
     key_val = "TEST_KEY12345678987654"
     key_name = "cloud_key"
     full_path = (Path('stats') / DEFAULT_USER_CONF)
@@ -67,8 +85,6 @@ def test_write_key_no_conf_file(tmp_directory, monkeypatch):
     key_val = "TEST_KEY12345678987654"
     key_name = "cloud_key"
     monkeypatch.setattr(telemetry, 'DEFAULT_HOME_DIR', '.')
-    stats = Path('stats')
-    stats.mkdir()
     full_path = (Path('stats') / DEFAULT_USER_CONF)
 
     # Write cloud key to existing file, assert on key/val
@@ -80,7 +96,7 @@ def test_write_key_no_conf_file(tmp_directory, monkeypatch):
     assert key_val in conf[key_name]
 
 
-def test_overwrites_api_key(write_sample_conf):
+def test_overwrites_api_key():
     key_val = "TEST_KEY12345678987654"
     key_name = "cloud_key"
     full_path = (Path('stats') / DEFAULT_USER_CONF)
@@ -97,15 +113,15 @@ def test_overwrites_api_key(write_sample_conf):
 
 
 @pytest.mark.parametrize('arg', [None, '12345'])
-def test_api_key_well_formatted(write_sample_conf, arg):
+def test_api_key_well_formatted(arg):
     with pytest.raises(BaseException) as excinfo:
         cloud.set_key(arg)
 
     assert 'The API key is malformed' in str(excinfo.value)
 
 
-def test_get_api_key(monkeypatch, write_sample_conf, capsys):
-    monkeypatch.delenv('PLOOMBER_CLOUD_KEY', raising=True)
+def test_get_api_key(monkeypatch, capsys):
+    monkeypatch.delenv('PLOOMBER_CLOUD_KEY', raising=False)
 
     key_val = "TEST_KEY12345678987654"
     runner = CliRunner()
@@ -130,8 +146,8 @@ def test_get_api_key_from_env_var(monkeypatch):
     assert key_val in result.stdout
 
 
-def test_get_no_key(monkeypatch, write_sample_conf, capsys):
-    monkeypatch.delenv('PLOOMBER_CLOUD_KEY', raising=True)
+def test_get_no_key(monkeypatch, capsys):
+    monkeypatch.delenv('PLOOMBER_CLOUD_KEY', raising=False)
 
     runner = CliRunner()
     result = runner.invoke(get_key, catch_exceptions=False)
@@ -139,8 +155,8 @@ def test_get_no_key(monkeypatch, write_sample_conf, capsys):
     assert 'No cloud API key was found.\n' == result.stdout
 
 
-def test_two_keys_not_supported(monkeypatch, write_sample_conf, capsys):
-    monkeypatch.delenv('PLOOMBER_CLOUD_KEY', raising=True)
+def test_two_keys_not_supported(monkeypatch, capsys):
+    monkeypatch.delenv('PLOOMBER_CLOUD_KEY', raising=False)
 
     key_val = "TEST_KEY12345678987654"
     key2 = 'SEC_KEY12345678987654'
@@ -159,7 +175,7 @@ def test_two_keys_not_supported(monkeypatch, write_sample_conf, capsys):
     assert key2 in res.stdout
 
 
-def test_cloud_user_tracked(write_sample_conf):
+def test_cloud_user_tracked():
     key_val = "TEST_KEY12345678987654"
     runner = CliRunner()
     runner.invoke(set_key, args=[key_val], catch_exceptions=False)
@@ -214,7 +230,9 @@ def test_write_pipeline_no_valid_key(monkeypatch):
     assert 'API_Key' in res
 
 
-def test_write_pipeline_no_status_id():
+def test_write_pipeline_no_status_id(monkeypatch):
+    monkeypatch.setenv('PLOOMBER_CLOUD_KEY', 'TEST_KEY12345678987654')
+
     pipeline_id = ''
     status = 'started'
     res = write_sample_pipeline(pipeline_id, status)
@@ -422,7 +440,6 @@ def test_email_conf_file(tmp_directory, monkeypatch):
     monkeypatch.setattr(telemetry, 'DEFAULT_HOME_DIR', '.')
 
     stats = Path('stats')
-    stats.mkdir()
     conf_path = stats / telemetry.DEFAULT_USER_CONF
     conf_path.write_text("sample_conf_key: True\n")
 
@@ -443,6 +460,7 @@ def test_email_write_only_once(tmp_directory, monkeypatch):
     assert not input_mock.called
 
 
+@pytest.mark.skip(reason="moved email prompt to onboarding workflow")
 def test_email_call_on_examples(tmp_directory, monkeypatch):
     email_mock = Mock()
     monkeypatch.setattr(examples, '_email_input', email_mock)
@@ -450,6 +468,7 @@ def test_email_call_on_examples(tmp_directory, monkeypatch):
     email_mock.assert_called_once()
 
 
+@pytest.mark.skip(reason="moved email prompt to onboarding workflow")
 @pytest.mark.parametrize('user_email', ['email@ploomber.io', ''])
 def test_email_called_once(tmp_directory, monkeypatch, user_email):
     monkeypatch.setattr(telemetry, 'DEFAULT_HOME_DIR', '.')